fix(typescript): add proper types
